version=11.9.36_dev # new feature: GenNBO file reader for xx.31-.41 files and .nbo output if in that directory # bug fix: Spartan and WebMO orbitals are already normalized!

+++ Jmol/src/org/jmol/adapter/readers/quantum/GaussianReader.java
@@ -284,7 +284,7 @@ public class GaussianReader extends MOReader {
       tokens = getTokens(); // get the tokens in the line
       Atom atom = atomSetCollection.addNewAtom();
       atom.elementNumber =
-        (byte)parseInt(tokens[STD_ORIENTATION_ATOMIC_NUMBER_OFFSET]);
+        (short)parseInt(tokens[STD_ORIENTATION_ATOMIC_NUMBER_OFFSET]);
       if (atom.elementNumber < 0)
         atom.elementNumber = 0; // dummy atoms have -1 -> 0
       int offset = tokens.length-3;
